BJS2-79195 fix magic num, added method validUser
